reorganize hooks and view model code based on our organization strategy
